+++ concordium-std/src/impls.rs
@@ -1,6 +1,14 @@
-use crate::{convert, mem, num, prims, prims::*, traits::*, types::*};
+use crate::{
+    collections::{BTreeMap, BTreeSet},
+    convert::{self, TryFrom, TryInto},
+    mem, num, prims,
+    prims::*,
+    traits::*,
+    types::*,
+    vec::Vec,
+    String,
+};
 use concordium_contracts_common::*;
-
 use mem::MaybeUninit;
 
 impl convert::From<()> for Reject {
@@ -88,7 +96,6 @@ impl Seek for ContractState {
     type Err = ();
 
     fn seek(&mut self, pos: SeekFrom) -> Result<u64, Self::Err> {
-        use core::convert::TryFrom;
         use SeekFrom::*;
         match pos {
             Start(offset) => match u32::try_from(offset) {
@@ -145,7 +152,6 @@ impl Seek for ContractState {
 
 impl Read for ContractState {
     fn read(&mut self, buf: &mut [u8]) -> ParseResult<usize> {
-        use core::convert::TryInto;
         let len: u32 = {
             match buf.len().try_into() {
                 Ok(v) => v,
@@ -204,7 +210,6 @@ impl Write for ContractState {
     type Err = ();
 
     fn write(&mut self, buf: &[u8]) -> Result<usize, Self::Err> {
-        use core::convert::TryInto;
         let len: u32 = {
             match buf.len().try_into() {
                 Ok(v) => v,
@@ -257,7 +262,6 @@ impl HasContractState<()> for ContractState {
 /// # Trait implementations for Parameter
 impl Read for Parameter {
     fn read(&mut self, buf: &mut [u8]) -> ParseResult<usize> {
-        use core::convert::TryInto;
         let len: u32 = {
             match buf.len().try_into() {
                 Ok(v) => v,
@@ -347,7 +351,6 @@ impl Iterator for PoliciesIterator {
             get_policy_section(buf.as_mut_ptr() as *mut u8, 2 + 4 + 8 + 8 + 2, self.pos);
             buf.assume_init()
         };
-        use convert::TryInto;
         let skip_part: [u8; 2] = buf[0..2].try_into().unwrap_abort();
         let ip_part: [u8; 4] = buf[2..2 + 4].try_into().unwrap_abort();
         let created_at_part: [u8; 8] = buf[2 + 4..2 + 4 + 8].try_into().unwrap_abort();
@@ -676,5 +679,113 @@ impl<A: fmt::Debug> ExpectNoneReport for Option<A> {
         if let Some(x) = self {
             crate::fail!("{}: {:?}", msg, x)
         }
+    }
+}
+
+impl<K: Serial + Ord> SerialCtx for BTreeSet<K> {
+    fn serial_ctx<W: Write>(
+        &self,
+        size_len: schema::SizeLength,
+        out: &mut W,
+    ) -> Result<(), W::Err> {
+        schema::serial_length(self.len(), size_len, out)?;
+        serial_set_no_length(self, out)
+    }
+}
+
+impl<K: Deserial + Ord + Copy> DeserialCtx for BTreeSet<K> {
+    fn deserial_ctx<R: Read>(
+        size_len: schema::SizeLength,
+        ensure_ordered: bool,
+        source: &mut R,
+    ) -> ParseResult<Self> {
+        let len = schema::deserial_length(source, size_len)?;
+        if ensure_ordered {
+            deserial_set_no_length(source, len)
+        } else {
+            deserial_set_no_length_no_order_check(source, len)
+        }
+    }
+}
+
+impl<K: Serial + Ord, V: Serial> SerialCtx for BTreeMap<K, V> {
+    fn serial_ctx<W: Write>(
+        &self,
+        size_len: schema::SizeLength,
+        out: &mut W,
+    ) -> Result<(), W::Err> {
+        schema::serial_length(self.len(), size_len, out)?;
+        serial_map_no_length(self, out)
+    }
+}
+
+impl<K: Deserial + Ord + Copy, V: Deserial> DeserialCtx for BTreeMap<K, V> {
+    fn deserial_ctx<R: Read>(
+        size_len: schema::SizeLength,
+        ensure_ordered: bool,
+        source: &mut R,
+    ) -> ParseResult<Self> {
+        let len = schema::deserial_length(source, size_len)?;
+        if ensure_ordered {
+            deserial_map_no_length(source, len)
+        } else {
+            deserial_map_no_length_no_order_check(source, len)
+        }
+    }
+}
+
+impl<T: Serial> SerialCtx for &[T] {
+    fn serial_ctx<W: Write>(
+        &self,
+        size_len: schema::SizeLength,
+        out: &mut W,
+    ) -> Result<(), W::Err> {
+        schema::serial_length(self.len(), size_len, out)?;
+        serial_vector_no_length(self, out)
+    }
+}
+
+impl<T: Deserial> DeserialCtx for Vec<T> {
+    fn deserial_ctx<R: Read>(
+        size_len: schema::SizeLength,
+        _ensure_ordered: bool,
+        source: &mut R,
+    ) -> ParseResult<Self> {
+        let len = schema::deserial_length(source, size_len)?;
+        deserial_vector_no_length(source, len)
+    }
+}
+
+impl SerialCtx for &str {
+    fn serial_ctx<W: Write>(
+        &self,
+        size_len: schema::SizeLength,
+        out: &mut W,
+    ) -> Result<(), W::Err> {
+        schema::serial_length(self.len(), size_len, out)?;
+        serial_vector_no_length(&self.as_bytes().to_vec(), out)
+    }
+}
+
+impl SerialCtx for String {
+    fn serial_ctx<W: Write>(
+        &self,
+        size_len: schema::SizeLength,
+        out: &mut W,
+    ) -> Result<(), W::Err> {
+        self.as_str().serial_ctx(size_len, out)
+    }
+}
+
+impl DeserialCtx for String {
+    fn deserial_ctx<R: Read>(
+        size_len: schema::SizeLength,
+        _ensure_ordered: bool,
+        source: &mut R,
+    ) -> ParseResult<Self> {
+        let len = schema::deserial_length(source, size_len)?;
+        let bytes = deserial_vector_no_length(source, len)?;
+        let res = String::from_utf8(bytes).map_err(|_| ParseError::default())?;
+        Ok(res)
     }
 }